confirm password and forget password

+++ src/app/home/home.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { first } from 'rxjs/operators';
-import { UserService, AuthenticationService, AlertService, courseCreaterService } from '../_services';
+import { UserService, AuthenticationService, AlertService } from '../_services';
 import { ActivatedRoute, Router } from '@angular/router';
 import Swal from 'sweetalert2'
 @Component({ templateUrl: 'home.component.html' })
@@ -11,34 +11,58 @@ export class HomeComponent implements OnInit {
     email: string;
     userDetails: any = [];
     constructor(
-        private authenticationService: AuthenticationService, private router: Router, public coursecreaterService: courseCreaterService,
+        private authenticationService: AuthenticationService, private router: Router,
         private userService: UserService, private route: ActivatedRoute, private alertService: AlertService
     ) {
         this.route.queryParams.subscribe(params => {
             let email = params['email'];
             if (email) {
                 this.router.navigate(['/dummy'])
-                var user = localStorage.getItem('userDetails');
-                this.userDetails = JSON.parse(user);
-                this.userService.updateProfile(email, this.userDetails._id)
-                    .pipe(first())
-                    .subscribe(
-                        data => {
-                            if (data.data['updateUser'].data.success === true) {
-                                localStorage.setItem('currentUser', 'true');
-                                this.router.navigate(['/home']);
-                            } else {
+                if (params['type'] == 'registration') {
+                    var user = localStorage.getItem('userDetails');
+                    this.userDetails = JSON.parse(user);
+                    this.userService.updateProfileStatus(email, this.userDetails._id)
+                        .pipe(first())
+                        .subscribe(
+                            data => {
+                                if (data.data['updateUser'].data.success === true) {
+                                    localStorage.setItem('currentUser', 'true');
+                                    this.router.navigate(['/home']);
+                                } else {
+                                    localStorage.setItem('currentUser', null)
+                                    this.router.navigate(['/dummy'])
+                                    this.router.navigate(['/LMSlogin']);
+                                    Swal.fire('Failed',data.data['updateUser'].data.message,'error');
+                                }
+                            },
+                            error => {
                                 localStorage.setItem('currentUser', null)
-                                this.router.navigate(['/dummy'])
                                 this.router.navigate(['/LMSlogin']);
-                                Swal.fire(data.data['updateUser'].data.message);
-                            }
-                        },
-                        error => {
-                            localStorage.setItem('currentUser', null)
-                            this.router.navigate(['/LMSlogin']);
-                            Swal.fire('Please try after sometime');
-                        });
+                                Swal.fire('Please try after sometime','','error');
+                            });
+                } else if (params['type'] == 'forgotpassword') {
+                    this.authenticationService.checklinkstatus(email)
+                        .pipe(first())
+                        .subscribe(
+                            data => {
+                                if (data.data['checklinkstatus'].success === true) {
+                                    console.log(data)
+                                    this.router.navigate(['/CMSconfirmPassword', {email: email }]);
+                                } else {
+                                    console.log(data)
+                                    localStorage.setItem('currentUser', null)
+                                    this.router.navigate(['/dummy'])
+                                    this.router.navigate(['/']);
+                                    Swal.fire('Failed',data.data['checklinkstatus'].message,'error');
+                                }
+                            },
+                            error => {
+                                localStorage.setItem('currentUser', null)
+                                this.router.navigate(['/']);
+                                Swal.fire('Please try after sometime','','error');
+                            });
+                }
+
             }
 
         });
@@ -49,18 +73,7 @@ export class HomeComponent implements OnInit {
         this.email = this.route.snapshot.paramMap.get('email');
     }
 
-    // deleteUser(id: number) {
-    //     this.userService.delete(id)
-    //         .pipe(first())
-    //         .subscribe(() => this.loadAllCourses());
-    // }
-
     private loadAllCourses() {
-        this.coursecreaterService.getCourse()
-            .pipe(first())
-            .subscribe((users: any) => {
-                this.users = users;
-                // console.log(this.users)
-            });
+
     }
 }
